Implement user-specific plugin installation
This commit implements the oft-requested user-specific plugin
installation. This resolves the pull requests: #2, #4, #8.

Notable changes:

Add a test cookbook that implements two scenarios of plugin
installation. The first is for a user that will exist in a test-kitchen
environment - vagrant. The second is for a user created during the
chef run itself, to prove we can look up the user and get the right
home directory.

A new `user` attribute, which is a "global" default if using the
plugins attribute to install a list of plugins. This assumes a single
user will have vagrant plugins. Consumers of this cookbook with
multiple users that have vagrant installations on a single system will
need to implement their own recipes.

Add a helper method, `vagrant_get_home`, that attempts to be smart
about detecting the user for plugin installation.

Refactor the provider actions to use `execute` resources (and add
`use_inline_resources`). This gives us the benefit of built-in
`converge_by`, and thus why-run support. It also makes the action
blocks easier to read.

Move the plugin installation loop out of the default recipe and into
its own recipe. This gives consumers flexibility, and decouples the
functionality.