Closes #31, return on unsupported platforms
The attributes files in every cookbook in a Chef dependency set are
evaluated, because Chef doesn't know what the actual reason for the
dependency is. As such, when this cookbook is used in a transitive
dependency as described in #31, the lookup methods that return
attribute values based on platform will bail out when the platform is
not supported. Rather than introduce branching conditionals around
those attributes, it is cleaner to simply "return."

This commit also adds clear language to the README about the platform
requirements being in line with the packages available to download
from the Vagrant site.